Add Grid.Ratio mode

+++ src/fe_image.cpp
@@ -2410,12 +2410,12 @@ FeRectangle *FeImage::add_rectangle(float x, float y, float w, float h)
 	return NULL;
 }
 
-FeImage *FeImage::add_surface(int w, int h)
+FeImage *FeImage::add_surface(float w, float h)
 {
 	return add_surface( 0, 0, w, h );
 }
 
-FeImage *FeImage::add_surface(float x, float y, int w, int h)
+FeImage *FeImage::add_surface(float x, float y, float w, float h)
 {
 	FePresentableParent *p = m_tex->get_presentable_parent();
 	if ( p )
